requirements.txt
Django==3.1
django-environ==0.4.5
django-extensions==3.0.5
gunicorn==20.0.4
Pillow==7.2.0
psycopg2==2.8.5
django-mptt==0.11.0
shortuuid==1.0.1
djangorestframework==3.11.1
django-filter==2.4.0
